# api.quiz/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz Platform API",
        "description": "API documentation for Quiz Platform with User and Role management",
        "contact": {
            "email": "admin@quizplatform.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.quiz.test",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/courses/{course}/contents": {
            "get": {
                "tags": [
                    "Course Contents"
                ],
                "summary": "List contents",
                "operationId": "6fa55995c46f0ee234dede48667043e5",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Course Contents"
                ],
                "summary": "Create content",
                "operationId": "5adaf431f87aaad74b9b6e681b3d8430",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "title"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "pdf",
                                            "text",
                                            "video",
                                            "quiz",
                                            "certificate"
                                        ]
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "order_index": {
                                        "type": "integer"
                                    },
                                    "duration_seconds": {
                                        "type": "integer"
                                    },
                                    "payload": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/courses/{course}/contents/{content}": {
            "put": {
                "tags": [
                    "Course Contents"
                ],
                "summary": "Update content",
                "operationId": "b702273763c7e5d41422094ec0026e81",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "content",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Course Contents"
                ],
                "summary": "Delete content",
                "operationId": "d1ed33198fd115f8c257aebed355fb45",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "content",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}